requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.22
asyncpg==0.27.0
alembic==1.11.1
pydantic==2.6.2
pydantic-settings==2.2.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
croniter==1.4.1
python-dotenv==1.1.1
httpx==0.25.0
pytest==7.4.0
pytest-asyncio==0.22.0
pytest-httpx==0.25.0
python-multipart==0.0.6
loguru==0.7.0
typing-extensions==4.9.0
email-validator==2.1.0.post1
